+++ core/mod.js
@@ -3,4 +3,5 @@ export * from "./fetch-router.js";
 export * from "./http.js";
 export * from "./migrator.js";
 export * from "./postgres.js";
+export * from "./structures.js";
 export * from "./utilities.js";